# ws-consumer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing real-time messaging",
        "title": "Real-Time Messaging Consumer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_inbound_rest_v1_handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/real-time-messaging_consumer_pkg_http.SuccessResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/real-time-messaging_consumer_pkg_http.ErrorResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/real-time-messaging_consumer_pkg_http.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is healthy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/real-time-messaging_consumer_pkg_http.SuccessResponseData"
                        }
                    }
                }
            }
        },
        "/v1/ws": {
            "get": {
                "description": "This endpoint establishes a WebSocket connection but cannot be tested via Swagger UI. Use a WebSocket client instead",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket Connection",
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "x-hidden": true
            }
        }
    },
    "definitions": {
        "internal_adapters_inbound_rest_v1_handlers.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "real-time-messaging_consumer_pkg_http.ErrorResponseData": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "description": "\"error\"",
                    "type": "string"
                }
            }
        },
        "real-time-messaging_consumer_pkg_http.SuccessResponseData": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "description": "\"success\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}